move MyJIT to separate file
